Remove forced cast to Number in Plural Rules Functions
The input value for ResolvePlural should not be cast to Number, as doing
so will lose the trailing zeros when the input is originally a string.
Dropping the cast and the assertion will not cause any regressions, as
the remaining `isFinite(x)` check in ResolvePlural will fail for any
string that cannot be parsed as a StringNumericLiteral.